Creer structure de rafles

+++ rafle/calc/rafle_tree_calc.c
@@ -1,4 +1,5 @@
 #include "rafle_tree_calc.h"
+#include <SDL2/SDL_stdinc.h>
 #include <stdio.h>
 
 void spitOut(pawn pawns[], pawn NPawns[], Case damier[NB_CASE_LG][NB_CASE_LG], int indEater, int iEater, int jEater, int indVictim, int add0, int add1){
@@ -54,7 +55,7 @@ PathTree* divideAndGather(pawn pawns[],pawn NPawns[], Case damier[NB_CASE_LG][NB
     int di, dj;
     PathTree* res = pathTreeCreateNode(i, j);
     PathTree* child;
-    for (int k = 0; k < 4; k++) {
+    for (Uint8 k = 0; k < 4; k++) {
         //Petite astuce pour parcourir les voisins :)
         getDirsFromCode(k, &di, &dj);
         if (!eatingIsOutOfBounds(i, j, di, dj) && canEat(pawns, damier, indEater, i, j, di, dj)) {
@@ -67,7 +68,7 @@ PathTree* divideAndGather(pawn pawns[],pawn NPawns[], Case damier[NB_CASE_LG][NB
     //C'est de l'eugenisme, donc ca reste dans le theme des rafles.
     int maxDepth = pathTreeDepth(res) - 1;
     int depth;
-    for (int k = 0; k < 4; k++) {
+    for (Uint8 k = 0; k < 4; k++) {
         getDirsFromCode(k, &di, &dj);
 
         PathTree* inspected = pathTreeChild(res, dj, di);
